+++ recordfile/test.txt
@@ -1,4 +1,4 @@
-﻿数字索引	字符串索引	数字类型	字符串类型	数组类型	嵌套数组	变长数组	结构体类型
-1	one	0	knife	"[1, 2]"	"[[0,1], [1,2], [2,3]]"	"[1, 2, 3]"	"{""name"": ""name5566"", ""num"": 1}"
-2	two	0	cat	"[3, 4]"	"[[1,2], [2,3], [3,4]]"	"[4, 5]"	"{""name"": ""name5566"", ""num"": 2}"
-3	three	0	book	"[5, 6]"	"[[2,3], [3,4], [4,5]]"	[6]	"{""name"": ""name5566"", ""num"": 3}"
+﻿数字索引	字符串索引	数字类型	字符串类型	数组类型	嵌套数组	变长数组	结构体类型	Map类型
+1	one	0	knife	"[1, 2]"	"[[0,1], [1,2], [2,3]]"	"[1, 2, 3]"	"{""name"": ""name5566"", ""num"": 1}"	"{""Mysql1"": {""Host"":""127.0.0.1"",""Port"":""3306"",""User"":""root"",""Passwd"":""111"",""Db"":""db1""},""Mysql2"": {""Host"":""127.0.0.2"",""Port"":""3306"",""User"":""root"",""Passwd"":""aaa"",""Db"":""db2""}}"
+2	two	0	cat	"[3, 4]"	"[[1,2], [2,3], [3,4]]"	"[4, 5]"	"{""name"": ""name5566"", ""num"": 2}"	"{""Mysql1"": {""Host"":""127.0.0.1"",""Port"":""3306"",""User"":""root"",""Passwd"":"""",""Db"":""db1""},""Mysql2"": {""Host"":""127.0.0.2"",""Port"":""3306"",""User"":""root"",""Passwd"":""bbb"",""Db"":""db2""}}"
+3	three	0	book	"[5, 6]"	"[[2,3], [3,4], [4,5]]"	[6]	"{""name"": ""name5566"", ""num"": 3}"	"{""Mysql1"": {""Host"":""127.0.0.1"",""Port"":""3306"",""User"":""root"",""Passwd"":"""",""Db"":""db1""},""Mysql2"": {""Host"":""127.0.0.2"",""Port"":""3306"",""User"":""root"",""Passwd"":""ccc"",""Db"":""db2""}}"
